refactor: reorganize types and constants into stripe-specific files (#105908)
