+++ lib/core/payment/stripe/constants.ts
@@ -1,6 +1,24 @@
+// ---- CORE IMPORTS ---- //
+import {BankTransferCurrency, BankTransferType, CountryCode} from './types';
+
 export const BANK_TRANSFER_STATUS = {
   PAID: 'paid',
   PENDING: 'pending',
+} as const;
+
+export const BANK_TRANSFER_TYPE = {
+  EU: 'eu_bank_transfer',
+  US: 'us_bank_transfer',
+} as const;
+
+export const BANK_TRANSFER_CURRENCY = {
+  EUR: 'eur',
+  USD: 'usd',
+} as const;
+
+export const BANK_ACCOUNT_TYPE = {
+  IBAN: 'iban',
+  ABA: 'aba',
 } as const;
 
 export const PAYMENT_INTENT_STATUS = {
@@ -27,3 +45,42 @@ export const STRIPE_PAYMENT_METHOD_TYPE = {
   CARD: 'card',
   CUSTOMER_BALANCE: 'customer_balance',
 } as const;
+
+export const BANK_TRANSFER_CONFIGS: Record<
+  BankTransferCurrency,
+  {
+    type: BankTransferType;
+    recommendedCountries: CountryCode[];
+    default: CountryCode;
+  }
+> = {
+  [BANK_TRANSFER_CURRENCY.EUR]: {
+    type: BANK_TRANSFER_TYPE.EU,
+    recommendedCountries: ['FR', 'DE', 'ES', 'IT', 'NL', 'BE'],
+    default: 'FR',
+  },
+  [BANK_TRANSFER_CURRENCY.USD]: {
+    type: BANK_TRANSFER_TYPE.US,
+    recommendedCountries: ['US'],
+    default: 'US',
+  },
+};
+
+export const COUNTRY_TO_BANK_TRANSFER: Record<
+  CountryCode,
+  {currency: BankTransferCurrency; type: BankTransferType}
+> = Object.entries(BANK_TRANSFER_CONFIGS).reduce(
+  (acc, [currency, config]) => {
+    config.recommendedCountries.forEach(country => {
+      acc[country] = {
+        currency: currency as BankTransferCurrency,
+        type: config.type,
+      };
+    });
+    return acc;
+  },
+  {} as Record<
+    CountryCode,
+    {currency: BankTransferCurrency; type: BankTransferType}
+  >,
+);
